Move crawler into support file
